Don't detect getrandom.
This fixes an issue where the configure script detects getrandom even
when it is not present (e.g. in a cross-compilation environment or on
Ubuntu 18.04).

getrandom is only used for GUID creation and the fallback function reads
sizeof(GUID) bytes from /dev/urandom, which exactly what getrandom is
defined to do.